StudentsHandler initial push
