+++ src/main.py
@@ -7,9 +7,8 @@ import sys
 import argparse
 import os
 from datetime import datetime
-import load_csv
-import student_error_checker
 import pandas as pd
+import load_csv
 #import run_ga ...when GA is completed and tested
 #import scoring ...when Scoring is completed and tested
 
@@ -83,7 +82,7 @@ if __name__ == "__main__":
     # read, parse, and handle errors of all three csv files
     load_csv.settingsHandler(settingsFileData)
     load_csv.projectsHandler(projectsFileData)
-    student_error_checker.studentsHandler(studentsFile) 
+    load_csv.studentsHandler(studentsFile) 
 
     if progMode == 'Assignment':
         print("\nProgram running in Assignment mode with a max run time of {0} minutes.".format(load_csv.maxRunTime))
